fix(persistence-powerup): Changed size field in Image to long

+++ IMuseum.Persistence/Models/Image.cs
@@ -8,7 +8,7 @@ public record Image : DatabaseModel
     public string Title { get; set; }
     public byte[] Bytes { get; set; }
     public string FileExtension { get; set; }
-    public decimal Size { get; set; }
+    public long Size { get; set; }
     /// <summary>
     /// Artwork of this image. (Related from relation to artworks)
     /// </summary>
